requirements.txt
asgiref==3.4.1
certifi==2021.10.8
Django==3.2.5
django-environ==0.8.1
djangorestframework==3.13.1
pip==21.2.2
psycopg2==2.8.6
pytz==2021.3
setuptools==61.2.0
sqlparse==0.4.1
typing_extensions==4.1.1
wheel==0.37.1
wincertstore==0.2
